Implementando camada de segurança na aplicação

+++ src/main/java/com/br/medtech/service/DoctorService.java
@@ -20,7 +20,7 @@ public class DoctorService {
     @Autowired
     private UserAcessService userAcessService ;
 
-    public ResponseEntity<Optional<Doctor>> findDoctorById(Long id) {
+    public ResponseEntity<Optional<Doctor>> findDoctorById(Integer id) {
         try {
             Optional<Doctor> doctor = doctorRepository.findById(id);
 
@@ -46,7 +46,7 @@ public class DoctorService {
         }
     }
 
-    public ResponseEntity<Doctor> update(Long id, Doctor doctor) {
+    public ResponseEntity<Doctor> update(Integer id, Doctor doctor) {
         if (!doctorRepository.existsById(id)) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Médico não encontrado no banco de dados.");
         }
@@ -60,7 +60,7 @@ public class DoctorService {
         }
     }
 
-    public ResponseEntity<Doctor> delete(Long id){
+    public ResponseEntity<Doctor> delete(Integer id){
         if (!doctorRepository.existsById(id)) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Médico não encontrado no banco de dados.");
         }
