+++ src/main/java/com/br/medtech/service/PatienteService.java
@@ -19,7 +19,7 @@ public class PatienteService {
     @Autowired
     private UserAcessService userAcessService;
 
-    public ResponseEntity<Optional<Patient>> findPatientById(Long id) {
+    public ResponseEntity<Optional<Patient>> findPatientById(Integer id) {
         try {
             Optional<Patient> patient = patientRepository.findById(id);
 
@@ -45,7 +45,7 @@ public class PatienteService {
         }
     }
 
-    public ResponseEntity<Patient> update(Long id, Patient Patient) {
+    public ResponseEntity<Patient> update(Integer id, Patient Patient) {
         if (!patientRepository.existsById(id)) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Paciente não encontrado no banco de dados.");
         }
@@ -59,7 +59,7 @@ public class PatienteService {
         }
     }
 
-    public ResponseEntity<Patient> delete(Long id){
+    public ResponseEntity<Patient> delete(Integer id){
         if (!patientRepository.existsById(id)) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Paciente não encontrado no banco de dados.");
         }
